updated ej vectores

+++ 1. Java Básico/EjerciciosSemana1/Clase2EjerciciosPracticos/src/ejercicio3/Vectores.java
@@ -17,16 +17,16 @@ public class Vectores {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
 		
-		double[] tempMaximas = new double[7];
-		double	suma = 0;
+		int[] tempMaximas = new int[7];
+		int	suma = 0;
 		
 		for(int i = 0; i < tempMaximas.length; i++) {
 			System.out.println("Introduce una temperatura máxima");
-			tempMaximas[i] = sc.nextDouble();
+			tempMaximas[i] = sc.nextInt();
 			suma += tempMaximas[i];
 		}
 		
-		System.out.printf("El promedio es %2.f\n", (suma / tempMaximas.length));
+		System.out.printf("El promedio es %2.f\n", (double)(suma / tempMaximas.length));
 		
 		sc.close();
 	}
